requirements.txt
# external requirements
click
Sphinx
coverage
awscli>=1.27.130
flake8
python-dotenv>=0.5.1
geosyspy
fastapi
hypercorn[trio]==0.14.3
pandas
requests
requests-mock
scipy
shapely
rasterio
xarray
pyproj
matplotlib
boto3
azure-storage-blob